rename FlashFormatterInterface into FlashFormatterBase
